refactor(primitive): migrate to SFC
