--- packages/radix-vue/src/Primitive/Primitive.ts
@@ -1,58 +1,0 @@
-import { type PropType, defineComponent, h } from 'vue'
-import { Slot } from './Slot'
-
-export type AsTag =
-  | 'a'
-  | 'button'
-  | 'div'
-  | 'form'
-  | 'h2'
-  | 'h3'
-  | 'img'
-  | 'input'
-  | 'label'
-  | 'li'
-  | 'nav'
-  | 'ol'
-  | 'p'
-  | 'span'
-  | 'svg'
-  | 'ul'
-  | 'template'
-  // eslint-disable-next-line @typescript-eslint/ban-types
-  | ({} & string) // any other string
-
-export interface PrimitiveProps {
-  /**
-   * Setting "asChild" to true has the same effect as setting "as" to "template".
-   * @default false
-   */
-  asChild?: boolean
-  /**
-   * @default "div"
-   */
-  as?: AsTag
-}
-
-export const Primitive = defineComponent({
-  name: 'Primitive',
-  inheritAttrs: false,
-  props: {
-    asChild: {
-      type: Boolean,
-      default: false,
-    },
-    as: {
-      type: String as PropType<AsTag>,
-      default: 'div',
-    },
-  },
-  setup(props, { attrs, slots }) {
-    const asTag = props.asChild ? 'template' : props.as
-
-    if (asTag !== 'template')
-      return () => h(props.as, attrs, { default: slots.default })
-
-    return () => h(Slot, attrs, { default: slots.default })
-  },
-})
